--- packages/cw-orch-core/src/contract/paths/mod.rs
@@ -1,8 +1,0 @@
-pub use artifacts_dir::from_workspace;
-pub use artifacts_dir::ArtifactsDir;
-pub use github::{GithubWasmPath, GithubWasmPathLocation};
-pub use wasm_path::WasmPath;
-
-mod artifacts_dir;
-mod github;
-mod wasm_path;
